매치-마스터-배틀 동기화 작업 중 마스터 서버의 빈방 체크 조건 수정이 필요함

+++ MasterServer/MasterServer/LanServer.h
@@ -99,7 +99,7 @@ typedef struct st_BattleRoom
 	long BattleServerNo;
 	int RoomNo;
 	int MaxUser;
-	int CurUser;
+	long CurUser;
 	char EnterToken[32];
 	std::vector<RoomPlayerInfo> RoomPlayer;
 }BattleRoom;
